% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juantools.R
\name{as.num.fac}
\alias{as.num.fac}
\title{Convert factor to numeric}
\usage{
as.num.fac(x)
}
\arguments{
\item{x}{A factor}
}
\value{
x A numeric representation of x
}
\description{
Convert a factor to string then to numeric, using as.num.char
}
\examples{
as.num.fac(factor('$1,000.00'))

}
