% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juantools.R
\name{digit_sep}
\alias{digit_sep}
\title{Format thousands place separator}
\usage{
digit_sep(x, sep = " ")
}
\arguments{
\item{x}{A number}

\item{sep}{Separator; default is space, \code{" "}}
}
\value{
A number with correct separator
}
\description{
This function changes the printed thousands separator in a number,
using space as the default
}
\examples{
digit_sep(1000)
digit_sep(10000)

}
