% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juantools.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{Function to print nice scientific notation in plot legends}
\usage{
fancy_scientific(x)
}
\arguments{
\item{x}{Numeric}
}
\value{
Expression for x in plotmath scientific notation
}
\description{
This function formats a number to an expression in plotmath format
}
