% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juantools.R
\name{percent_sep}
\alias{percent_sep}
\title{Convert decimal to percent and add space before percent sign}
\usage{
percent_sep(x, places = 2, sep = TRUE)
}
\arguments{
\item{x}{Numeric}

\item{places}{Integer; how many places to round to?}

\item{sep}{Logical; add space before percent sign? Default is \code{TRUE}}
}
\value{
x String reprenting x as percent, with space added before
percent sign
}
\description{
Convert decimal to percent and add space before percent sign
}
\examples{
percent_sep(0.12345)
percent_sep(0.12345, places=3)
percent_sep(0.12345, sep=FALSE)

}
