% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juantools.R
\name{sqm2sqft}
\alias{sqm2sqft}
\title{Convert numeric square meters to square feet}
\usage{
sqm2sqft(x)
}
\arguments{
\item{x}{Numeric; assumed units are sq m}
}
\value{
x Numeric, converted to sq ft
}
\description{
Wrapper to sqft2sqm that converts in the reverse direction
}
\examples{
sqm2sqft(100)
# identical
sqft2sqm(100, reverse=TRUE)

}
