% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juantools.R
\name{thou_sep}
\alias{thou_sep}
\title{Format thousands place separator for numbers bigger than 9999}
\usage{
thou_sep(x, sep = " ")
}
\arguments{
\item{x}{A number}

\item{sep}{Separator; default is space, \code{" "}}
}
\value{
A number with correct separator
}
\description{
Wrapper to \code{digit_sep} to print thousands separated by space,
\emph{except} for numbers in the 1000s
}
\examples{
thou_sep(1000)
thou_sep(10000)

}
