% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juantools.R
\name{using}
\alias{using}
\title{Load packages, installing if not already installed}
\usage{
using(...)
}
\arguments{
\item{...}{package names, can be a list, a vector, or a bunch of
strings}
}
\description{
This function takes a list of package names and tries to load them;
any packages that are not found are installed
}
